😁(code_parser): add `code_parser` module and implement its integration with application
This commit introduces a new code_parser module. It allows the parsing of shape drawing operations from a string. This was necessary to replace hardcoded vector line generation with operations defined in code. This change gives flexibility in dynamically manipulating the shape to be drawn. Modified also included files to implement the usage of the new parser module.

The parser supports operations for move, line, quad, and cubic, handling whitespace and commas in between arguments. It provides a method to parse the entire code into a list of operations that can be further used by the VecLineGen.

In addition, the hardcoded drawing operations in the src/main.rs have been replaced by the parsed operations from the user-inputted code. Code samples in src/default_code file were also updated to test the functionality.

Signed-off-by: W-Mai <1341398182@qq.com>