😁(main): add error visualization on the plot, and enhance error text display
New eframe::egui::WidgetText uses were introduced to support the addition of a RichText component for improved error text display. An error is now added to the plot in dark red when it occurs during the computation process. This makes the errors more visible by providing immediate visual feedback directly on the plot. Furthermore, in case of successful computation, the plot line color was changed to dark blue for better visual consistency.

Signed-off-by: W-Mai <1341398182@qq.com>