😁(code_parser): add comment parsing functionality to CodeParser
Implemented functionality for parsing and ignoring comments in the input code. The changes allow CodeParser to differentiate single and multiline comments, incrementing the cursor position accordingly. This functionality ensures that comments are not processed as code. Also simplified code by replacing "self.curr_pos() < self.code.len()" with "self.not_eof()" as a means to check the cursor position against the length of the code.

Signed-off-by: W-Mai <1341398182@qq.com>