😁(vec_line_gen): update method to generate vector of vectors in vec_line_gen
The gen method in vec_line_gen has been updated to produce a vector of vectors. This change was necessary to cater for multiple plots to be added simultaneously. It helps increase the performance and efficiently manages memory. This change also required updates in main.rs where the plots are being drawn. Now, each vector from the list of vectors gets its color and line width.

Signed-off-by: W-Mai <1341398182@qq.com>