😍(main): add custom component for toggling LCD Coordinates
Implemented a custom toggle component for enabling and disabling LCD Coordinates feature. Updated main.rs to use this new component in the UI. When toggled on, the y-axis values are inverted to match the LCD coordinate system's behaviour. This new feature gives users the flexibility to switch between standard and LCD coordinate systems as per their preference.

Signed-off-by: W-Mai <1341398182@qq.com>