😁(main): add colorful block feature and optimize imports
In this commit, we added a new feature to enable colorful blocks in the visualization. This is achieved by introducing a COLOR_PALETTE constant and using it to color the blocks. A toggle option "Colorful Blocks" was also added in the configurations section to let users switch this feature on and off.

Additionally, we optimized the list of imports, specifically the `std::ops` imports by merging the `Add` and `RangeInclusive` into a single `use` statement.

Signed-off-by: W-Mai <1341398182@qq.com>