😁(main): fix y-axis format and layout improvements in visualization code
Changed the y-axis formatting in the plotting function from 2 decimal places to none for both lcd_coords conditions in order to improve the clarity of the graph. Reduced size of the builder from 100.0 to 60.0 in the strip function for a more balanced layout. Lastly, improved readability by centering the "Controls" heading. These changes improve visual readability and user interaction with the controls.

Signed-off-by: W-Mai <1341398182@qq.com>