⚙️(code_parser): remove redundant 'cursor_back' function in code_parser
The 'cursor_back' function was removed from the `code_parser/mod.rs` file. This function was not needed anymore, as the error handling has been improved. With these changes, the cursor does not need to be moved backward when an error occurs. Instead, an appropriate error with the current cursor position is returned. This simplifies the parser logic and makes the code cleaner and easier to maintain. We also reordered some lines of code to maintain coherence in the sequence of function calls.

Signed-off-by: W-Mai <1341398182@qq.com>