🚧(refactor): update vector operations and data visualization functionality
Updated the vector operations and data visualization functionality in various modules. The changes include the addition of VecLineData struct and implementation of IVisData trait for it. This change helps to better organize vector line data and their positions. Updated functions in `gen.rs` and `mod.rs` to include new functionality and adapt to these changes. Updated `interfaces.rs` to adjust to the changes in data types. Further changes were also made to 'main.rs' to compliment these alterations. The purpose is mainly to improve data handling and readability of code.

Signed-off-by: W-Mai <1341398182@qq.com>