🚧(refactor): remove specific VecOps enums and refactor VecLineGen for general commands
Replaced VecOps enums in VecLineGen with Command<f64> for a more generic approach of handling vector operations. This refactoring of code eliminates the duplication of adding and parsing specific VecOps by introducing a more generic command type. This change increases code reusability and makes the function definitions more concise and readable. Removed specific add functions for VecOps as it is no longer needed.

Signed-off-by: W-Mai <1341398182@qq.com>