🚧(matrix): add matrix transformation to interfaces and VecLineData
Implemented a matrix transformation function in the base interface and the VecLineData struct. This allows for transformation of vectors using a 3x3 matrix. This was done to add flexibility in performing translations, scaling, and rotations directly on vector data. This will centralize transformation operations, making the code easier to manage and read. Changes in 'main.rs' reflect initial testing of the new functionality with an identity matrix.

Signed-off-by: W-Mai <1341398182@qq.com>