🚧(refactor): add TextDataSrc and update parser to utilize it
This commit introduces TextDataSrc struct in 'data_src.rs' that wraps around data source and importantly uses 'Rc<RefCell>' allowing 'CodeParser' to borrow its contents mutably multiple times.

Additionally, 'parser.rs' is updated to use TextDataSrc as a parameter for 'new()' and 'parse()'. Making 'parse' to be a mutable method allows the parser instance to be reused for subsequent parses.

Finally, the addition of 'data_src' module is exposed in 'mod.rs' and interfaces for 'IDataSource' and 'IParser' have been updated to reflect the changes.

These changes are mainly aimed to increase efficiency and allow the data source to be shared across multiple parsers. Changes on 'main.rs' were made to adhere to the updated interfaces.

Signed-off-by: W-Mai <1341398182@qq.com>