🚧(refactor): improve vector plotting with new CommonVecVisualizer
This commit introduces the CommonVecVisualizer struct for better plotting of vectors. This struct implements the IVisualizer trait for added flexibility. It features an improved plot function in main.rs that deals better with LCD coordinates and transformation matrices. The visualizer provides a more readable, maintainable and efficient solution for vector visualization. The IVisualizer trait in interfaces.rs has also been updated to support the new struct and its functions.

Signed-off-by: W-Mai <1341398182@qq.com>