🐟(default): add Sakaban Turtle object to default code
The source file, default_code, has been updated to include the drawing and movements for Sakaban Turtle. This introduces a new object to our existing set. The new object was added below the existing code, without any alteration to the existing structure. This addition will enrich the visual graphics of our program.

Signed-off-by: W-Mai <1341398182@qq.com>