🚧(refactor): refactor interfaces with associated types for cleaner use
This commit replaces concrete types in interfaces (IParser, IVisDataGenerator, IVisualizer, ICodeEditor) with associated types aiming to reduce the complexity and improve the flexibility of these traits. This refactoring follows the principle of 'program to an interface, not an implementation', and allows us to apply different types for different implementations without restructuring the code. Consequently, all the implementing classes have been updated to match the new interfaces' definitions. This gives more flexibility to future development and can easily extend or adapt to new variations.

Signed-off-by: W-Mai <1341398182@qq.com>