🚧(Deps):refactored code to improve compatibility and fix dependencies
This commit comprises multiple changes principally aimed to enhance the project's compatibility and fix dependencies issues. A redundant, fixed type associated with the `ICodeEditor` trait was removed and replaced with a more flexible dynamic reference to `ICommandSyntax`. This change provides more flexibility for using different command syntaxes. In tandem with this, any corresponding changes in the `show` method of various modules were adjusted accordingly. In addition, all dependencies in `Cargo.toml` were updated to their latest compatible versions to resolve any outstanding issues and enable usage of the latest features. Finally, depreciated methods of `Plot` in `Visualizer` and `gen` modules were replaced with the current ones.

Signed-off-by: Benign X <1341398182@qq.com>