🚧(AnyData): add AnyData struct to manage dynamic types
This change introduces a struct `AnyData` in a new file `any_data.rs`. The `AnyData` struct will serve as an abstract data type and provide methods for casting values to wanted types. It also modifies other parts of the codebase (`interfaces.rs`, `gen.rs`, and `parser.rs`), wherever possible, to use this new `AnyData` in argument vectors, thereby enabling more generic operations. The main reason for this change is to improve flexibility and ease the process of adding new features in the future due to the generic nature of `AnyData`.

Signed-off-by: Benign X <1341398182@qq.com>