🚧(IVisData): add data method to IVisData in interfaces.rs
A new data method has been added in IVisData interface in the file interfaces.rs in order to extend the flexibility and functionality of the type. With this method, `AnyData` struct is now being used instead of a numeric type opening up the possibilities of a more generic data type manipulation. This is a crucial refactoring aimed at enhancing adaptability for future extensions.

Signed-off-by: Benign X <1341398182@qq.com>