🚧(WIP): refactor and adjust vector operations
The code makes modifications to how vectors are generated and visualized. For vector generation, the code now creates a clone of each vector, which allows for grouping. For visualization, an additional check has been applied for non-empty point vector. Overall, these changes aim to improve efficacy and functionality of vector operations.

Signed-off-by: Benign X <1341398182@qq.com>