❤️(WPA): support web run mode
This commit updates the dependencies in Cargo.toml and segregates UI-related code and application logic in separate modules. It also adds the capability to render the application in web browsers as a Progressive Web App (PWA), alongside its native version. Changes involve setting up a Service Worker for offline caching and creating an HTML file as a rendering target. Finally, various image assets and a Web Manifest are added for displaying the app on different platforms.

Signed-off-by: Benign X <1341398182@qq.com>