😁(Cargo): add flag for web-sys unstable APIs in cargo config
Added a configuration line for target wasm32-unknown-unknown with a rustflag to enable web_sys_unstable_apis. This is necessary as clipboard API in web-sys is currently unstable, to make copy (ctrl + c) work. The link for checking the status and a reference to a related issue in rust analyzer build cache invalidation is included in a comment.

Signed-off-by: Benign X <1341398182@qq.com>