😊(UI): Add samples panel and simple square sample
Added a new "Samples" side panel to the UI, which showcases various code samples including a simple square. This enhancement introduces a visual representation of each sample, allowing for better user understanding and interaction. The `config_fn` parameter was also added to relevant functions to enable customization of the plot settings.

Signed-off-by: Benign X <1341398182@qq.com>